requirements.txt
ninja
transformers==4.46.0
accelerate==1.0.1
bitsandbytes==0.43.3
datasets==3.0.0
peft==0.13.0
trl==0.11.0
sentence-transformers==3.1.1
protobuf==3.20.*
openai==1.53.0
claudette==0.1.0
vllm==0.6.4.post1
hydra-core
sentencepiece
einops
kagglehub
wandb